renamer en del .tsx-filer uten JSX til .ts
