+++ src/sider/08-vedlegg/form/variants/belopEn/belopEnFormToGenericOkonomiInput.ts
