+++ src/sider/08-vedlegg/form/variants/bruttoNetto/bruttoNettoFormToLonnsInput.ts
